requirements.txt
# Core App Framework
gradio==5.49.1

# LangChain Core
langchain==1.0.7
langchain-core==1.0.5
langchain-text-splitters==1.0.0

# LangChain Integrations
langchain-chroma==1.0.0
langchain-huggingface==1.0.1
langchain-openai

# Vector Database
chromadb==1.3.4

# HuggingFace & Embeddings
huggingface-hub==0.36.0
sentence-transformers>=3.0.0
transformers>=4.0

# Document Processing
pypdf2==3.0.1
python-docx==1.2.0

# Environment & Config
python-dotenv==1.2.1